requirements.txt
pdfplumber>=0.10.0
pandas>=2.0.0
openpyxl>=3.1.2
flask>=3.0.0
flask-cors>=4.0.0
PyPDF2>=3.0.1
werkzeug>=3.0.1
numpy>=1.26.0
gunicorn>=22.0.0
pycryptodome>=3.19.0
sqlalchemy
pymysql
python-dotenv>=1.0.0